+++ tests/test_bpy_handlers_rst.py
@@ -19,6 +19,7 @@
 # THE SOFTWARE.
 
 
+from __future__ import unicode_literals
 from docutils import nodes
 from docutils.parsers.rst import Directive
 
@@ -135,3 +136,7 @@ class HandlerTestCase(test_base.BaseHandlerTestCase):
   # =====
 
   test_generate_str_EXPECT = '<p>\xc3\xa1</p>'
+
+  # =====
+
+  test_smartypants_EXPECT = '<p>foo &#8220;bar&#8221;</p>'